Interner Server Error On Create Profile Screen Api

+++ lib/model/Product Card View Model/porductCardListData.dart
@@ -10,7 +10,7 @@ class ProductCardListData {
   String? discountPrice;
   String? image;
   int? stock;
-  double? star;
+  int? star;
   String? remark;
   int? categoryId;
   int? brandId;
